refactor: Move TransferOptimizationPreferences to preferences
